requirements.txt
attrs==22.1.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.21.3
filelock==3.8.0
Flask==2.2.3
html5lib==1.1
idna==3.4
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
MarkupSafe==2.1.2
more-itertools==8.14.0
msgpack==1.0.4
packaging==22.0
pexpect==4.8.0
pipenv==2023.3.20
pkginfo==1.9.6
platformdirs==2.5.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.18.1
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
urllib3==1.26.12
virtualenv==20.21.0
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.2.3

python-dotenv~=1.0.0
PyJWT~=2.6.0
marshmallow~=3.19.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
